requirements.txt
aws-cdk-lib>=2.126.0
aws-cdk.aws-scheduler-alpha>=2.126.0a0
aws-cdk.aws-scheduler-targets-alpha>=2.126.0a0
cdk-nag>=2.28.154
constructs>=10.0.0,<11.0.0
boto3>=1.29.4
jsii>=1.94.0
